python: force usage of python 3.10 from unstable as pyopenssl workaround
note that this does not actually fix the issue, but i've opted to retain
the unstable tracking in case a fix is available sooner. `pyopenssl` is
a very common dependency, so it being broken on `aarch64-darwin` is
difficult to deal with. for now, i've just disabled support for that
platform since i'm not using it (though i expect that might change soon)

cf. https://github.com/NixOS/nixpkgs/issues/175875