fix(darwin)!: remove yabai and skhd
BREAKING CHANGES:

I just don't have the patience to deal with circumventing macOS'
anti-user lockdown of the window manager. `yabai` is great, but it requires too
much maintenance. And when macOS inevitably breaks something, I am put out of
commission for at least a day due to a total reliance on `yabai` working. So, I
figure it'd be better to avoid reverse-engineered Good Tools in favor of the
shiny by-the-book paid tools that work within the walled garden, just the way
Apple wants their ecosystem to work.