fix(workstation|kitty): include kitty in workstation feature group
i've fiddled with wezterm recently but i think i'm still much more a fan of
kitty "just working" the way i would expect -- probably largely because i've
used it for so long and have it configured extensively.

i also find kitty's documentation to be way more digestable and less fragmented
than wezterm's. it is so amazingly difficult to get a sense of the available
themes in wezterm (even though it bundles, like, all of the themes in the world)
because the docs ui is not well-suited to its subject (e.g. indexed alphanumerically
over 27 pages, with some of those pages containing dozens of themes while others
are almost empty, no builtin way to preview themes).