fix(emacs|nixpkgs): ensure latest emacs packages are available
with a major caveat (see comment) -- requires applying the overlay globally due
to 1) our approach to using a single nixpkgs instance globally (where possible)
but moreso 2) the emacs-overlay does not seem to provide any way of accessing
these packages outside of the overlay.

perhaps there's a way to fake it by passing dummy `pkgs` to the overlay
function(s)... but at that point, fuck it, that's enough evidence that it's
probably wiser to let emacs handle its own packages. there are drawbacks to that
approach too, of course -- some packages have special dependencies or need extra
hand-holding for environment integration, which nixpkgs handles as needed (not
the overlay, which just provides updated sources).