fix(tuvok): restore default nixpkgs to facilitate update ease
As the updated DisplayLink driver has still not been merged upstream,
this commit also reverts to the older version of the DisplayLink driver.  I
initially sought out the update to alleviate an unusable frame-dropping, but I
think that likely has more to do with the recent changes to the graphics module,
which nixos-apple-silicon does not yet account for.