fix(firefox): unmanage addons
The benefits outweigh the drawbacks of letting Firefox retain full
control over extension management:

- If I don’t run `nixos-rebuild` for a while, then extensions aren’t
updated.  This causes problems especially with 1Password, which wants to
be up to date for security reasons.
- Managing extensions with Nix requires that Firefox Sync be disabled
for extensions.  I actually like Firefox Sync because it… keeps things
in sync.  That’s especially useful on Android and Windows, which are not
managed by Nix.
- Managed extensions require packaging and an entire CI infrastructure
for keeping that updated.