global: pin nixpkgs to nixpkgs-apple-silicon
This will make it a lot easier to keep hosts in sync.
tuuvok is prone to build failures due to upstream Nixpkgs
breakages, so it is safest for that host to track the
Asahi-specific Nixpkgs version.  But that Nixpkgs is just
another Nixpkgs -- nothing different from any other Nixpkgs
input aside from its tracked commit.

One example of the type of breakage that can happen is that
some newer version of Nixpkgs might remove an option, which
requires a change to configuration.  If that happens to be
configuration shared by all hosts, then it will need to be
a global change.  But if some hosts do not know about the
new option, everything falls apart and there is no easy
way forward.