keyboard: ensure wheel can use `uinput`
This was, by coincidence, a lurking oversight.  The NixOS module for
kanata automatically handles enabling the `uinput` kernel module, which
is required to use kanata as well as espanso.  When I switched to a
user-specific home module for kanata and disabled the NixOS-level kanata
support, `uinput` went away.  However, I did not immediately notice
until after a reboot, because the module was still loaded.

As a side note, I am not sure how espanso was working without my user
being in the `uinput` group.